requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# Data Processing
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Hugging Face
huggingface_hub>=0.16.0

# Utilities
python-multipart>=0.0.6
python-json-logger>=2.0.7
psutil>=5.9.0
